backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Pipecat — real-time voice/video AI pipeline (requires Python 3.10+)
pipecat-ai[google,silero]>=0.0.108
